build: reject implicit engine size narrowing on MSVC

+++ engine/botlib/l_script.cpp
@@ -748,7 +748,7 @@ static int PS_ReadPunctuation(script_t *script, token_t *token)
 		punc = &script->punctuations[i];
 #endif //PUNCTABLE
 		p = punc->p;
-		len = strlen(p);
+		len = (int)( strlen(p) );
 		//if the script contains at least as much characters as the punctuation
 		if (script->script_p + len <= script->end_p)
 		{
