+++ engine/renderer/tr_mesh.cpp
@@ -186,7 +186,7 @@ int R_ComputeLOD( trRefEntity_t *ent ) {
 		{
 			int frameSize;
 			mdr = (mdrHeader_t *) tr.currentModel->modelData;
-			frameSize = (size_t) (&((mdrFrame_t *)0)->bones[mdr->numBones]);
+			frameSize = (int)( (size_t) (&((mdrFrame_t *)0)->bones[mdr->numBones]) );
 			
 			mdrframe = (mdrFrame_t *) ((byte *) mdr + mdr->ofsFrames + frameSize * ent->e.frame);
